requirements.txt
opencv-python
fer
moviepy
tensorflow
numpy
pyserial
